rearranged the logic of containers and renamed them to be more semantic, fetched API
