+++ client/containers/MonthContainer.jsx
@@ -1,13 +1,15 @@
 import React from 'react';
+import WeekContainer from './WeekContainer.jsx';
 
 const MonthContainer = () => {
 
     return(
-        <div className='monthOuterContainer'>
+   
         <div className='monthContainer'>
-             Month here
+             <h2>Month container here</h2>
+             <WeekContainer/>
         </div>
-      </div>
+
     )
 }
