+++ client/containers/WeatherContainer.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Weather from '../components/weather.jsx'
+
+const WeatherContainer = () => {
+
+    return(
+        <div className='weatherContainer'>
+             <h3>weather container here</h3>
+             <Weather/>
+        </div>
+    )
+}
+
+export default WeatherContainer;
