--- client/containers/day.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const DayContainer = () => {
-
-    return(
-        <div className='dayOuterContainer'>
-        <div className='dayContainer'>
-             Day here
-        </div>
-      </div>
-    )
-}
-
-export default DayContainer;
