+++ client/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { createRoot } from 'react-dom/client';
 import App from './App.jsx';
+import styles from './stylesheets/styles.scss'
 
 
 const root = createRoot(document.getElementById('root'));
